+++ CHANGE_LOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Attributes are now passed through the logger's attribute formatter before serializing them to JSON for inclusion in job payloads.
 - **Breaking Change** Updated terminology to match Lumberjack 2 change to refer to "attributes" rather than "tags".
   - Changed config option `log_tag_prefix` to `log_attributes_prefix`
   - Renamed `Lumberjack::Sidekiq::TagPassthroughMiddleware` to `Lumberjack::Sidekiq::AttributePassthroughMiddleware`.